refactor: Cards -> Deck 변경, Deck 싱글톤 패턴 적용
