+++ src/test/java/blackjack/domain/player/UserDeckTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.List;
 import java.util.stream.Stream;
 
-class UserCardsTest {
+class UserDeckTest {
 
     private static Stream<Arguments> provideCards() {
         return Stream.of(
